hunyuan3d 添加render_npu 光栅化结果复用 优化开关代码



# message auto-generated for no-merge-commit merge:
!20 merge master into master

hunyuan3d 添加render_npu 光栅化结果复用 优化开关代码

Created-by: chenjingzzzz
Commit-by: chenjingzzzz
Merged-by: turing_project1
Description: # Pull Request 模板
----

## 描述
请提供此Pull Request的预期功能，以方便检视。

## 类型
- [ ] Bug 修复
- [ ] 新功能
- [ ] 重构（即不是新增功能，也不是修改bug的代码变动）
- [ ] 构建过程或辅助工具的变动
- [ ] 文档内容更新

## 如何测试
描述测试这个变更的步骤，包括哪些文件需要被修改。

## Checklist:
- [ ] 我的代码遵循这个项目的代码风格
- [ ] 我已经自己测试过我的代码
- [ ] 我已经更新了相应的文档
- [ ] 我已经在标题中正确使用了类型标签（例如：`feat`, `fix`, `refactor`, `docs`, `test`）

## 其他信息
在这里可以添加任何与这个 Pull Request 相关的其他说明。


See merge request: cann/cann-recipes-spatial-intelligence!20

+++ models/Hunyuan3D/hy3dgen/texgen/utils/dehighlight_utils.py
@@ -85,11 +85,11 @@ class Light_Shadow_Remover():
             image_array[:, :, rgb_channel] = alpha_channel
             image = Image.fromarray(image_array)
 
-            image_tensor = torch.tensor(np.array(image) / 255.0).to(self.device)
+            image_tensor = torch.tensor(np.array(image) / 255.0).to(torch.float16).to(self.device)
             alpha = image_tensor[:, :, rgb_channel:]
             rgb_target = image_tensor[:, :, :rgb_channel]
         else:
-            image_tensor = torch.tensor(np.array(image) / 255.0).to(self.device)
+            image_tensor = torch.tensor(np.array(image) / 255.0).to(torch.float16).to(self.device)
             alpha = torch.ones_like(image_tensor)[:, :, :1]
             rgb_target = image_tensor[:, :, :rgb_channel]
 
@@ -106,7 +106,7 @@ class Light_Shadow_Remover():
             guidance_scale=self.cfg_text,
         ).images[0]
 
-        image_tensor = torch.tensor(np.array(image) / 255.0).to(self.device)
+        image_tensor = torch.tensor(np.array(image) / 255.0).to(torch.float16).to(self.device)
         rgb_src = image_tensor[:, :, :rgb_channel]
 
         image = self.recorrect_rgb(rgb_src, rgb_target, alpha)
